additional backtests with other exposures and symbols

+++ backtest_custom.py
@@ -12,33 +12,51 @@ symbols = ["GALUSDT", "SOLUSDT", "TRXUSDT",
            "TOMOUSDT", "OCEANUSDT", "IMXUSDT",
            "ETHUSDT", "BTCUSDT", "ADAUSDT", "XRPUSDT", "XMRUSDT"]
 symbol_str = ",".join(symbols)
-long_exposure = np.arange(21) * 0.05
+long_exposure = np.arange(11) * 0.10
 long_exposure = long_exposure[1:]
-t0 = time.time()
-results_dict = {}
-for lw in long_exposure:
-    short_exposure = np.arange(5) * 0.25 * lw
-    for sw in short_exposure:
-        print(f"Elapsed time since start: {round(time.time() - t0, 2)} s")
-        print(f"Starting (lw, sw) = ({lw:.3f}, {sw:.3f})")
-        call_str = [
-            "python", "backtest.py", config_params,
-            "-b", config_backtest,
-            "-s", symbol_str,
-            "-oh", "yes",
-            "-lw", str(round(lw, 3)),
-            "-sw", str(round(sw, 3)),
-            # "-start_date", "2021-01-01",
-            # "-end_date", "2023-07-14"
-        ]
+if os.path.isfile('results_backtest_custom.pkl'):
+    with open('results_backtest_custom.pkl', 'rb') as fp:
+        results_dict = pickle.load(fp)
+else:
+    t0 = time.time()
+    results_dict = {}
+    for lw in long_exposure:
+        short_exposure = np.arange(4) * 0.333 * lw
+        for sw in short_exposure:
+            print(f"Elapsed time since start: {round(time.time() - t0, 2)} s")
+            print(f"Starting (lw, sw) = ({lw:.3f}, {sw:.3f})")
+            call_str = [
+                "python", "backtest.py", config_params,
+                "-b", config_backtest,
+                "-s", symbol_str,
+                "-oh", "yes",
+                "-lw", str(round(lw, 3)),
+                "-sw", str(round(sw, 3)),
+                # "-start_date", "2022-05-15",
+                # "-end_date", "2023-07-26"
+            ]
 
-        # Subprocess.run
-        results_i = subprocess.run(call_str, capture_output=True, text=True)
-        results_dict[(lw, sw)] = results_i
-        print(f"Finished (lw, sw) = ({lw:.3f}, {sw:.3f})")
-        with open('results_backtest_custom.pkl', 'wb') as fp:
-            pickle.dump(results_dict, fp, protocol=pickle.HIGHEST_PROTOCOL)
+            # Subprocess.run
+            results_i = subprocess.run(call_str, capture_output=True, text=True)
+            results_dict[(round(lw, 3), round(sw, 3))] = results_i
+            print(f"Finished (lw, sw) = ({lw:.3f}, {sw:.3f})")
+
+            with open('results_backtest_custom.pkl', 'wb') as fp:
+                pickle.dump(results_dict, fp, protocol=pickle.HIGHEST_PROTOCOL)
         # https://docs.python.org/3/library/subprocess.html
     # break
+
+# Content analysis
+full_text = ""
+for k, v in results_dict.items():
+    text = v.stdout
+    text = text.replace("\x1b[39m", "")
+    text = text.replace("\x1b[0m", "")
+    text = text.replace("\x1b[31m", "")
+    full_text += f"KEY: {k}\n\n" + text + "\n"
+
+text_file = open("full_report.txt", "wt")
+n = text_file.write(full_text)
+text_file.close()
 
 print("End of script!")
